Add ability to logout/close session
Currently there isn't a way to close a session once we are finished.
This little bit of cleanup will allow us to close a session when
switching to another vhost or just as a clean way to exit. We will also
take care of clearing our cached clone ticket since it will no longer be
valid.